fix: cast hidden states to float32 before contrastive differencing
fp16 subtraction between close activation vectors loses precision,
producing near-zero diff matrices that trigger LAPACK SLASCL errors
and SVD convergence failures.  Casting to float32 before the
subtraction preserves the signal.

Also fixes torch_dtype deprecation warning in _load_text_from_multimodal.

+++ liahona/model.py
@@ -109,7 +109,7 @@ def _load_text_from_multimodal(
     from safetensors.torch import load_file
     from transformers.utils import cached_file, SAFE_WEIGHTS_INDEX_NAME
 
-    model = AutoModelForCausalLM.from_config(text_config, torch_dtype=dtype)
+    model = AutoModelForCausalLM.from_config(text_config, dtype=dtype)
 
     index_path = cached_file(model_id, SAFE_WEIGHTS_INDEX_NAME)
     model_dir = os.path.dirname(index_path)
